+++ tests/testthat/test-parse.R
@@ -5,10 +5,6 @@ test_that("parse_quo() etc return quosures", {
   expect_identical(parse_quos("foo(bar)\n mtcars", "base"), list(set_env(quo(foo(bar)), base_env()), set_env(quo(mtcars), base_env())))
 })
 
-test_that("parse_expr() requires scalar character", {
-  expect_error(parse_expr(letters), "`x` must be a string or a R connection")
-})
-
 test_that("parse_quosure() and parse_quosures() are deprecated", {
   with_verbose_retirement({
     expect_warning(parse_quosure("foo"), "soft-deprecated")
@@ -28,4 +24,17 @@ test_that("temporary connections are closed", {
 test_that("parse_expr() throws meaningful error messages", {
   expect_error(parse_expr(""), "No expression to parse")
   expect_error(parse_expr("foo; bar"), "More than one expression parsed")
+})
+
+test_that("parse_exprs() and parse_quos() handle character vectors", {
+  exprs <- parse_exprs(c("foo; bar", "baz"))
+  attributes(exprs) <- NULL # For srcrefs
+  expect_identical(exprs, unname(exprs(foo, bar, baz)))
+
+  quos <- parse_quos(c("foo; bar", "baz"), current_env())
+  expect_identical(quos, quos(foo, bar, baz))
+})
+
+test_that("parse_exprs() requires connections or character vectors", {
+  expect_error(parse_exprs(env()), "must be a character vector or an R connection")
 })